+++ src/Adsefid.Sdk/Messenger/Models/SendP2PMessenger.cs
@@ -60,7 +60,7 @@ public sealed class SendP2PMessengerResponse
 
     /// <summary>Total cost charged across all receptors in this send.</summary>
     [JsonPropertyName("total_cost")]
-    public required long TotalCost { get; init; }
+    public required decimal TotalCost { get; init; }
 
     /// <summary>Breakdown of receptor counts by outcome.</summary>
     [JsonPropertyName("counts")]
@@ -98,5 +98,5 @@ public sealed class P2PMessengerReceptorResult
     public required int Status { get; init; }
 
     [JsonPropertyName("cost")]
-    public required long Cost { get; init; }
+    public required decimal Cost { get; init; }
 }
